+++ historical_data.py
@@ -13,6 +13,7 @@ add rollbacks and commits
 """
 
 # libs
+# import stockdwh_functions
 import requests
 import mysql.connector  # https://dev.mysql.com/doc/connector-python/en/
 import json
@@ -39,7 +40,7 @@ db_settings_table_name = sql_db_conn["db_settings_table_name"]  # settings table
 # open db connection
 cursor = cnxn.cursor()
 
-cursor.execute("SELECT download_settings_id, market, tick_interval, stock_type, stock_exchange, api_range_to_overwrite, download_interval_sec, daily_update_from_files, monthly_update_from_files FROM " + db_schema_name + "." + db_settings_table_name + "  order by start_download_ux_timestamp asc limit 1")
+cursor.execute("SELECT download_settings_id, market, tick_interval, stock_type, stock_exchange, api_range_to_overwrite, download_interval_sec, daily_update_from_files, monthly_update_from_files FROM " + db_schema_name + "." + db_settings_table_name + " WHERE daily_update_from_files = 1 order by start_download_ux_timestamp asc limit 1")
 download_setting = cursor.fetchall()
 download_settings_id = download_setting[0][0]
 market = download_setting[0][1]
@@ -52,21 +53,6 @@ daily_update_from_files = download_setting[0][7]
 monthly_update_from_files = download_setting[0][8]
 
 # todo: function get single setting to download: curr, daily, monthly
-
-def get_settings():
-    cursor = cnxn.cursor()
-
-    cursor.execute( "SELECT download_settings_id, market, tick_interval, stock_type, stock_exchange, api_range_to_overwrite, download_interval_sec, daily_update_from_files, monthly_update_from_files FROM " + db_schema_name + "." + db_settings_table_name + "  order by start_download_ux_timestamp asc limit 1")
-    download_setting = cursor.fetchall()
-    download_settings_id = download_setting[0][0]
-    market = download_setting[0][1]
-    tick_interval = download_setting[0][2]
-    stock_type = download_setting[0][3]
-    stock_exchange = download_setting[0][4]
-    range_to_download = download_setting[0][5]
-    download_interval_sec = download_setting[0][6]
-
-
 
 DAILY_HISTORY_DAYS = 3
 
@@ -90,7 +76,7 @@ TYPE = 'klines'
 APP_PATH = "tmp/"
 
 
-def get_files():
+def get_daily_files():
     # todo: file import
     r = requests.get(BASE_URL)
     open(FILE_PATH + ".zip", "wb").write(r.content)
@@ -117,22 +103,24 @@ def get_files():
     for i in csv_data:
         cursor.execute("INSERT INTO " + db_schema_name+"."+db_table_name +"(open_time, open, high, low, close, volume, close_time, quote_asset_volume, number_of_trades, taker_buy_base_asset_volume, taker_buy_quote_asset_volume, `ignore`, market, tick_interval, stock_type, stock_exchange) values (%s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s, %s)", (i[0], i[1], i[2], i[3], i[4], i[5], i[6], i[7], i[8], i[9], i[10], i[11], market, tick_interval, stock_type, stock_exchange))
         print(i)
-    cnxn.commit()
-    cursor.close()
-    cnxn.close()
+
     print("insert done")
 
 
-
+   # start_download_ux_timestamp = 1
+   # start_download_ux_timestamp
 
 
 FILENAME_LIST = get_filenames_to_download_daily()
 
-FILENAME = FILENAME_LIST[0]
-FILE_PATH = APP_PATH + FILENAME
-BASE_URL = "https://data.binance.vision/data/"+stock_type+"/"+FILE_LENGTH+"/"+TYPE+"/"+market+"/"+tick_interval+"/"+FILENAME+".zip"
 
-print(BASE_URL)
+for j in FILENAME_LIST:
+    FILENAME = FILENAME_LIST[0]
+    FILE_PATH = APP_PATH + FILENAME
+    BASE_URL = "https://data.binance.vision/data/"+stock_type+"/"+FILE_LENGTH+"/"+TYPE+"/"+market+"/"+tick_interval+"/"+FILENAME+".zip"
+    get_daily_files()
 
-print(FILE_PATH)
-get_files()
+
+cnxn.commit()
+cursor.close()
+cnxn.close()
