+++ test2.py
@@ -19,4 +19,3 @@ def get_filenames_to_download():
 print(get_filenames_to_download())
 
 
-1
